# backend/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Card List API Contract
  version: '0.2.0'
paths:
  /api/home:
    get:
      summary: Get the home dashboard snapshot for the top page
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomeDashboardResponse'
        '503':
          description: Temporary failure while resolving home dashboard data
          content:
            application/json:
              schema:
                type: object
                required:
                  - error
                properties:
                  error:
                    type: string
                    enum:
                      - home_temporary_failure
  /api/tags:
    get:
      summary: Search selectable tag options for the shared filter modal
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Search text for narrowing tag candidates
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/FilterOption'
  /api/collections:
    get:
      summary: Search selectable collection options for the shared filter modal
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Search text for narrowing collection candidates
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/FilterOption'
  /api/cards:
    get:
      summary: List non-archived cards with cursor pagination, filter state, and optional answer text
      parameters:
        - in: query
          name: cursor
          schema:
            type: string
          description: Cursor token (opaque, from previous response)
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: q
          schema:
            type: string
          description: Full-text search query for title/content/answer/tag related matches
        - in: query
          name: tagIds
          schema:
            type: string
          description: Comma-separated tag ids
        - in: query
          name: collectionIds
          schema:
            type: string
          description: Comma-separated collection ids
        - in: query
          name: filter
          schema:
            type: string
            enum: [today, overdue, unlearned]
          description: Single-select status filter chosen from the dropdown
        - in: query
          name: sort
          schema:
            type: string
            enum: [next_review_at, proficiency, created_at]
          description: Sort key for stable ordered pagination
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Card'
                  nextCursor:
                    type: string
                    description: opaque cursor to fetch next page
        '400':
          description: Invalid query or cursor/sort mismatch
    post:
      summary: Create a new study card
      description: |
        Creates a new study card from title, content, optional free-form tag names,
        and an optional existing collection selection. Unknown tag names are created
        before linking. Collection creation is out of scope for this endpoint.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '200':
          description: Card created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCardResponse'
        '400':
          description: Invalid request payload or unknown collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCardError'
        '500':
          description: Failed to persist card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCardError'
  /api/cards/bulk:
    post:
      summary: Execute a bulk action on selected cards
      description: |
        `archive` removes affected cards from the default list after a successful refresh.
        `addTags` and `removeTags` are idempotent: adding an already attached tag or removing a missing tag is treated as a successful no-op.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [archive, delete, addTags, removeTags]
                cardIds:
                  type: array
                  items:
                    type: string
                tagIds:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: OK
        '400':
          description: Invalid request payload
  /api/review/start:
    post:
      summary: Create a review session from the current card list filter
      description: |
        Resolves the current filter on the server and creates a review session from the
        ordered result set. At most 200 reviewable cards are included in a session.
        Any excluded cards are summarized in `snapshot.filterSummary.targetResolution`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartReviewRequest'
      responses:
        '200':
          description: Review session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartReviewResponse'
        '400':
          description: Invalid request body
        '404':
          description: No reviewable cards remained for the requested filter
        '503':
          description: Temporary backend failure before a session snapshot could be created; the client should offer retry or return to the list
  /api/review/sessions/{sessionId}:
    get:
      summary: Retrieve the latest snapshot for an existing review session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Current review session snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewSessionSnapshot'
        '404':
          description: Session not found
        '503':
          description: Temporary backend failure while loading the latest snapshot; the client may show the last cached snapshot and offer retry
  /api/review/sessions/{sessionId}/assessment:
    put:
      summary: Save or replace the assessment for the current card
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssessmentRequest'
      responses:
        '200':
          description: Updated snapshot after assessment save
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewSessionSnapshot'
        '404':
          description: Session not found
        '409':
          description: Card is not editable in the current session state
        '503':
          description: Temporary backend failure while saving the assessment; the client may keep the pending assessment locally and retry
  /api/review/sessions/{sessionId}/navigation:
    post:
      summary: Move to the previous or next card in the review session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NavigateReviewRequest'
      responses:
        '200':
          description: Updated snapshot after navigation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewSessionSnapshot'
        '404':
          description: Session not found
        '409':
          description: Next navigation requires an assessment on the current card
        '503':
          description: Temporary backend failure while navigating; the client may continue showing the last cached snapshot and offer retry
components:
  schemas:
    HomeDashboardResponse:
      type: object
      required:
        - generatedAt
        - summary
        - recentActivities
        - state
      properties:
        generatedAt:
          type: string
          format: date-time
        summary:
          $ref: '#/components/schemas/HomeSummary'
        recentActivities:
          type: array
          maxItems: 3
          items:
            $ref: '#/components/schemas/HomeRecentActivity'
        state:
          $ref: '#/components/schemas/HomeViewState'
    HomeSummary:
      type: object
      required:
        - todayDueCount
        - overdueCount
        - unlearnedCount
        - streakDays
      properties:
        todayDueCount:
          type: integer
          minimum: 0
        overdueCount:
          type: integer
          minimum: 0
        unlearnedCount:
          type: integer
          minimum: 0
        streakDays:
          type: integer
          minimum: 0
    HomeRecentActivity:
      type: object
      required:
        - id
        - type
        - occurredAt
        - label
      properties:
        id:
          type: string
        type:
          type: string
          enum: [review_completed, review_started, card_created]
        occurredAt:
          type: string
          format: date-time
        label:
          type: string
        count:
          type: integer
          minimum: 0
          nullable: true
    HomeViewState:
      type: object
      required:
        - firstUse
        - noReviewToday
      properties:
        firstUse:
          type: boolean
        noReviewToday:
          type: boolean
    Card:
      type: object
      required:
        - id
        - title
        - content
        - tags
        - proficiency
        - nextReviewAt
        - lastCorrectRate
        - isArchived
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        answer:
          type: string
          nullable: true
          description: Optional multiline plain-text answer associated with the card
        tags:
          type: array
          items:
            type: string
        collectionId:
          type: string
          format: uuid
          nullable: true
        proficiency:
          type: number
        nextReviewAt:
          type: string
          format: date-time
        lastCorrectRate:
          type: number
        isArchived:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    FilterOption:
      type: object
      required:
        - id
        - label
      properties:
        id:
          type: string
        label:
          type: string
        matchedBy:
          type: string
          nullable: true
          enum: [name, alias, null]
    CardListFilter:
      type: object
      properties:
        q:
          type: string
        filter:
          type: string
          enum: [today, overdue, unlearned]
        sort:
          type: string
          enum: [next_review_at, proficiency, created_at]
        tagIds:
          type: array
          items:
            type: string
        collectionIds:
          type: array
          items:
            type: string
    StartReviewRequest:
      type: object
      properties:
        filter:
          $ref: '#/components/schemas/CardListFilter'
        cardIds:
          type: array
          items:
            type: string
      anyOf:
        - required: [filter]
        - required: [cardIds]
    StartReviewResponse:
      type: object
      required:
        - snapshot
      properties:
        snapshot:
          $ref: '#/components/schemas/ReviewSessionSnapshot'
    ReviewSessionSnapshot:
      type: object
      required:
        - sessionId
        - status
        - currentIndex
        - totalCount
        - remainingCount
        - canGoPrev
        - canGoNext
        - filterSummary
        - summary
      properties:
        sessionId:
          type: string
          format: uuid
        status:
          type: string
          enum: [in_progress, completed]
        currentIndex:
          type: integer
          minimum: 0
        totalCount:
          type: integer
          minimum: 0
        remainingCount:
          type: integer
          minimum: 0
        canGoPrev:
          type: boolean
        canGoNext:
          type: boolean
        filterSummary:
          $ref: '#/components/schemas/ReviewFilterSummary'
        currentCard:
          allOf:
            - $ref: '#/components/schemas/ReviewCardSnapshot'
          nullable: true
        summary:
          $ref: '#/components/schemas/ReviewSessionSummary'
    ReviewCardSnapshot:
      type: object
      required:
        - cardId
        - title
        - content
        - tags
        - reviewReason
        - locked
      properties:
        cardId:
          type: string
        title:
          type: string
        content:
          type: string
        answer:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        collectionLabel:
          type: string
          nullable: true
        nextReviewAt:
          type: string
          format: date-time
        reviewReason:
          $ref: '#/components/schemas/ReviewReason'
        currentAssessment:
          $ref: '#/components/schemas/ReviewAssessment'
        locked:
          type: boolean
    ReviewReason:
      type: object
      required:
        - label
      properties:
        label:
          type: string
          description: Short explanation of why the current card is included in this review session
        detail:
          type: string
          nullable: true
          description: Optional supporting detail such as due time or matching filter condition
        source:
          type: string
          enum: [next_review_at, overdue, filter_match, unlearned, manual_selection]
          nullable: true
    ReviewFilterSummary:
      type: object
      required:
        - sort
        - tagLabels
        - collectionLabels
      properties:
        q:
          type: string
          nullable: true
        filter:
          type: string
          nullable: true
          enum: [today, overdue, unlearned, null]
        sort:
          type: string
          enum: [next_review_at, proficiency, created_at]
        tagLabels:
          type: array
          items:
            type: string
        collectionLabels:
          type: array
          items:
            type: string
        targetResolution:
          $ref: '#/components/schemas/ReviewTargetResolution'
    ReviewTargetResolution:
      type: object
      required:
        - matchedCount
        - includedCount
        - excludedCount
        - exclusionBreakdown
      properties:
        matchedCount:
          type: integer
          minimum: 0
        includedCount:
          type: integer
          minimum: 0
          maximum: 200
        excludedCount:
          type: integer
          minimum: 0
        exclusionBreakdown:
          type: array
          items:
            $ref: '#/components/schemas/ReviewTargetExclusion'
    ReviewTargetExclusion:
      type: object
      required:
        - reason
        - count
      properties:
        reason:
          type: string
          enum: [over_limit, unavailable]
        count:
          type: integer
          minimum: 1
    ReviewSessionSummary:
      type: object
      required:
        - forgotCount
        - uncertainCount
        - rememberedCount
        - perfectCount
        - assessedCount
        - totalCount
      properties:
        forgotCount:
          type: integer
          minimum: 0
        uncertainCount:
          type: integer
          minimum: 0
        rememberedCount:
          type: integer
          minimum: 0
        perfectCount:
          type: integer
          minimum: 0
        assessedCount:
          type: integer
          minimum: 0
        totalCount:
          type: integer
          minimum: 0
    ReviewAssessment:
      type: string
      nullable: true
      enum: [forgot, uncertain, remembered, perfect, null]
    UpdateAssessmentRequest:
      type: object
      required:
        - cardId
        - assessment
      properties:
        cardId:
          type: string
        assessment:
          type: string
          enum: [forgot, uncertain, remembered, perfect]
    NavigateReviewRequest:
      type: object
      required:
        - direction
      properties:
        direction:
          type: string
          enum: [prev, next]
    CreateCardRequest:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        tagNames:
          type: array
          items:
            type: string
            minLength: 1
        collectionId:
          type: string
          nullable: true
    CreateCardResponse:
      type: object
      required:
        - ok
        - card
      properties:
        ok:
          type: boolean
          enum: [true]
        card:
          $ref: '#/components/schemas/Card'
    CreateCardError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          enum: [invalid_body, bad_request, database_error]
        message:
          type: string
          nullable: true
        details:
          type: object
          nullable: true